refactor(inference): extract ollama tests to sibling file
primer-inference/src/ollama.rs had a single inline trailing test module
(`mod tests`@333) containing two nested submodules
(`classify_ollama_tests`, `reasoning_wiring`). Move the whole module into
a sibling file `src/ollama/tests.rs` and replace the inline block with
`#[cfg(test)] mod tests;`.

587 -> 333 lines (-254, now well under the ~500 guideline). The test
module is a descendant, so `use super::*` (both the outer module and the
nested submodules) resolves to `ollama` verbatim -- no visibility edits,
zero churn. All 199 primer-inference lib tests preserved at their exact
module paths.

Pure test-only relocation: no behaviour change, no incidental fixes.

cargo test -p primer-inference --lib: 199 passed
cargo clippy -p primer-inference --all-targets -- -D warnings: clean
cargo test --workspace: 51 ok blocks, 0 failures
cargo fmt --all -- --check: clean
